+++ src/Toml/Bi.hs
@@ -15,13 +15,7 @@ Reexports functions under @Toml.Bi.*@.
 -}
 
 module Toml.Bi
-       ( module Toml.Bi.Code
-       , module Toml.Bi.Combinators
-       , module Toml.Bi.Map
-       , module Toml.Bi.Monad
+       ( module Toml.Bi.Map
        ) where
 
-import Toml.Bi.Code
-import Toml.Bi.Combinators
 import Toml.Bi.Map
-import Toml.Bi.Monad